% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anara-package.R
\docType{package}
\name{anara-package}
\alias{anara}
\alias{anara-package}
\title{anara: Highlight Problems In Survey Data}
\description{
More about what it does (maybe more than one line) Use four spaces when indenting paragraphs within the Description.
}
\author{
\strong{Maintainer}: Mayarí Montes de Oca \email{mmd486@nyu.edu}

Authors:
\itemize{
  \item Patrick Anker \email{psanker@nyu.edu} (\href{https://orcid.org/0000-0003-2302-0445}{ORCID})
}

Other contributors:
\itemize{
  \item Global TIES for Children [copyright holder]
}

}
\keyword{internal}
