% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixing.R
\name{apply_fixes}
\alias{apply_fixes}
\title{Apply fixes to databases}
\usage{
apply_fixes(
  databases,
  fixes,
  unique_id_col,
  id_col,
  verbose = FALSE,
  force_fix = FALSE
)
}
\arguments{
\item{databases}{A named list of \code{data.frame} objects The list names correspond
to the entry in the \code{database} column of the fix file.}

\item{fixes}{A \code{data.frame} containing the \emph{verified} fixes. Must have
been through \code{anara::verify_fixes}.}

\item{unique_id_col}{The column name for the column that is the surrogate
key in all \code{databases}}

\item{id_col}{The column name for the column that is the primary key
in all \code{databases}}

\item{verbose}{Produces logging messages if \code{TRUE}}

\item{force_fix}{If \code{TRUE}, applies fixes identified as problematic in
\code{anara::verify_fixes}. Generally an unsafe operation!}
}
\value{
A list of corrected \code{data.frames}
}
\description{
Apply fixes to databases
}
