% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatting.R
\name{fix_format}
\alias{fix_format}
\title{Puts a dataset into the fix format}
\usage{
fix_format(
  dat,
  id_col,
  unique_id_col,
  database_col,
  var_pool,
  review_fields = c("problem", "verifier", "note"),
  edit_fields = c(what = "what", change_to = "change_to", change_from = "change_from")
)
}
\arguments{
\item{dat}{The result of \code{anara::verify_ids}}

\item{id_col}{The name of the ID, or primary key, column.}

\item{unique_id_col}{The name of the row ID, or surrogate key,
column.}

\item{database_col}{The name of the column that stores the database names}

\item{var_pool}{The pool of metric and issue columns}

\item{review_fields}{The names of fields to be used for verification}

\item{edit_fields}{The names of the fix columns}
}
\value{
\code{dat} in fix format
}
\description{
Generally only used from \code{anara::verify_ids}
}
