% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metric.R
\name{metric}
\alias{metric}
\title{Define a metric}
\usage{
metric(variable, func, tag)
}
\arguments{
\item{variable}{A character vector of variables to run this metric on}

\item{func}{A function that takes a single argument. This is evaluated
\emph{by} entity ID.}

\item{tag}{A string to be used as the issue ID in the output spreadseet
of \code{anara::verify_ids}}
}
\value{
A 'metric' object
}
\description{
A metric is a definition of some comparison. By default,
there are two "metrics": integer distance, and string similarity.
This tool expands that toolkit to arbitrary comparisons.
}
