% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metric.R
\name{metrics}
\alias{metrics}
\title{A collection of metrics}
\usage{
metrics(...)
}
\arguments{
\item{...}{\code{metric} objects}
}
\value{
An object of type 'metrics'
}
\description{
Primarily used in the \code{extra_metrics} parameter of
\code{anara::verify_ids}.
}
