% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verify.R
\name{verify_ids}
\alias{verify_ids}
\title{Verify record consistency across databases}
\usage{
verify_ids(
  dat_list,
  id_col,
  unique_id_col,
  file = NULL,
  database_col = "database",
  variables = NULL,
  extra_metrics = NULL,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{dat_list}{A named list of \code{data.frames}}

\item{id_col}{The name of the ID, or primary key, column.
For consistency, should be the same across datasets.}

\item{unique_id_col}{The name of the row ID, or surrogate key,
column. For consistency, should be the same across datasets.}

\item{file}{If not \code{NULL}, a path to where the output spreadsheet
will be saved.}

\item{database_col}{The column name to store the \code{dat_list} names}

\item{variables}{A character vector of integer or character columns
to be used for comparison across datasets.}

\item{extra_metrics}{A \code{metrics()} call that contains a collection of
\code{metric()} calls}

\item{verbose}{Enables logging}

\item{...}{Extra parameters passed to \code{anara::fix_format}}
}
\value{
A \code{data.frame} in the fix format
}
\description{
Compares demographic information across datasets to determine
if the entity identified with ID \code{x} is the same across all
datasets.
}
